scripts/requirements.txt
firebase-admin
ossapi
python-dotenv